feat: [factory-argument-signatures-prd-mergeability] - Restore pkg/cli/run size lint compliance
